pyproject.toml
[tool.poetry]
name = "brownie-simple-storage"
version = "0.1.0"
description = ""
authors = ["t0astt <t0astt93@gmail.com>"]
readme = "README.md"
packages = [{include = "brownie_simple_storage"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
eth-brownie = "^1.19.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

